refactor: extract SidebarItem
